tool: update build_sdk to not require ld script in every library

+++ build_sdk.py
@@ -736,10 +736,11 @@ def build_lib_component(
 
     link_script = Path(component_name) / "microkit.ld"
     dest = lib_dir / "microkit.ld"
-    dest.unlink(missing_ok=True)
-    shutil.copy(link_script, dest)
-    # Make output read-only
-    dest.chmod(0o744)
+    if link_script.exists():
+        dest.unlink(missing_ok=True)
+        shutil.copy(link_script, dest)
+        # Make output read-only
+        dest.chmod(0o744)
 
     include_dir = root_dir / "board" / board.name / config.name / "include"
     source_dir = Path(component_name) / "include"
